restored

+++ webproject/web/src/main/java/org.numisoft.webproject/controllers/CatalogController.java
@@ -1,7 +1,9 @@
 package org.numisoft.webproject.controllers;
 
+import org.apache.log4j.Logger;
 import org.numisoft.webproject.dto.BanknoteDto;
 import org.numisoft.webproject.services.BanknoteService;
+import org.numisoft.webproject.utils.Constants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
@@ -17,12 +19,15 @@ import org.springframework.web.bind.annotation.RequestParam;
 @Controller
 public class CatalogController {
 
+    Logger logger = Logger.getLogger(CatalogController.class);
+
     @Autowired
     BanknoteService banknoteService;
 
     @Autowired
     BanknoteDto banknoteDto;
 
+    /* Catalog set-up */
     @RequestMapping(value = "/catalog")
     public String showCatalog(ModelMap modelMap,
                               @RequestParam(value = "page", defaultValue = "1") Integer page) {
@@ -35,25 +40,39 @@ public class CatalogController {
         return "catalog";
     }
 
-    @RequestMapping(value = "/delete")
+    /* Admin deletes banknote from Catalog */
+    @RequestMapping(value = "/catalog/delete")
     public String removeFromCatalog(@RequestParam(value = "id", defaultValue = "0") Integer banknote_id) {
 
+        if (banknote_id == 0) {
+            return Constants.REDIRECT_TO_400_ERROR;
+        }
+
         banknoteService.removeBanknoteFromCatalog(banknote_id);
-        return "redirect:/catalog";
+
+        logger.debug("<<@>> Admin deleted banknote: id=" + banknote_id);
+
+        return Constants.REDIRECT_TO_CATALOG;
     }
 
+    /* Admin adds new banknote to Catalog */
+    @RequestMapping(value = "/catalog/add")
+    public String addToCatalog(@ModelAttribute BanknoteDto banknoteDto) {
 
-    @RequestMapping(value = "/add")
-    public String addToCatalog(ModelMap modelMap, @ModelAttribute BanknoteDto banknoteDto) {
+        if (banknoteDto == null) {
+            return Constants.REDIRECT_TO_400_ERROR;
+        }
 
-        String title = ((BanknoteDto) modelMap.get("banknoteDto")).getTitle();
+        String title = banknoteDto.getTitle();
         int nominal = banknoteDto.getNominal();
         String country = banknoteDto.getCountry();
         String link = banknoteDto.getLink();
 
         banknoteService.addBanknoteToCatalog(title, nominal, country, link);
 
-        return "redirect:/catalog";
+        logger.debug("<<@>> Admin added banknote: " + nominal + " " + title);
+
+        return Constants.REDIRECT_TO_CATALOG;
     }
